requirements.txt
fake-bpy-module-latest==20241010
filelock==3.16.1
fsspec==2024.9.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
smplx==0.1.28
sympy==1.13.3
torch==2.4.1
typing_extensions==4.12.2
